requirements.txt
Django==3.2.8
numpy==1.19.5
Pillow==8.2.0
scikit-learn==0.24.2
